pyproject.toml
[project]
name = "wmput"
version = "0.1.0"
description = "Bot for Bluesky that adds watermarks to posted illustrations and replaces them with reposts."
authors = [
    {name = "sheep-snow",email = "user.email=71830342+sheep-snow@users.noreply.github.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "boto3 (>=1.36.21,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "atproto (>=0.0.58,<0.0.59)",
    "aws-cdk-lib (>=2.179.0,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.6"
pytest = "^8.3.4"

[tool.poetry]
package-mode = false